--- src/main/java/com/fluxtion/runtime/server/subscription/EventSource.java
@@ -1,25 +1,0 @@
-/*
- * SPDX-FileCopyrightText: © 2024 Gregory Higgins <greg.higgins@v12technology.com>
- * SPDX-License-Identifier: AGPL-3.0-only
- *
- */
-
-package com.fluxtion.runtime.server.subscription;
-
-import com.fluxtion.runtime.annotations.feature.Experimental;
-
-/**
- * Publishes events to {@link EventToQueuePublisher}. Register an {@link EventSource} instance with {@link EventFlowManager}
- * to receive the target queue via the setEventToQueuePublisher callback method.
- *
- * @param <T>
- */
-@Experimental
-public interface EventSource<T> {
-
-    void subscribe(EventSubscriptionKey<T> eventSourceKey);
-
-    void unSubscribe(EventSubscriptionKey<T> eventSourceKey);
-
-    void setEventToQueuePublisher(EventToQueuePublisher<T> targetQueue);
-}
