+++ src/main/java/com/fluxtion/server/dutycycle/ComposingEventProcessorAgent.java
@@ -1,0 +1,137 @@
+/*
+ * SPDX-FileCopyrightText: © 2024 Gregory Higgins <greg.higgins@v12technology.com>
+ * SPDX-License-Identifier: AGPL-3.0-only
+ *
+ */
+
+package com.fluxtion.server.dutycycle;
+
+import com.fluxtion.agrona.concurrent.DynamicCompositeAgent;
+import com.fluxtion.agrona.concurrent.OneToOneConcurrentArrayQueue;
+import com.fluxtion.runtime.StaticEventProcessor;
+import com.fluxtion.runtime.annotations.feature.Experimental;
+import com.fluxtion.runtime.input.EventFeed;
+import com.fluxtion.runtime.lifecycle.Lifecycle;
+import com.fluxtion.runtime.service.Service;
+import com.fluxtion.server.FluxtionServer;
+import com.fluxtion.server.service.scheduler.DeadWheelScheduler;
+import com.fluxtion.server.service.scheduler.SchedulerService;
+import com.fluxtion.server.subscription.EventFlowManager;
+import com.fluxtion.server.subscription.EventSubscriptionKey;
+import lombok.extern.java.Log;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.function.Supplier;
+
+/**
+ *
+ */
+@Experimental
+@Log
+public class ComposingEventProcessorAgent extends DynamicCompositeAgent implements EventFeed<com.fluxtion.server.subscription.EventSubscriptionKey<?>> {
+
+    private final com.fluxtion.server.subscription.EventFlowManager eventFlowManager;
+    private final ConcurrentHashMap<String, Service<?>> registeredServices;
+    private final ConcurrentHashMap<com.fluxtion.server.subscription.EventSubscriptionKey<?>, EventQueueToEventProcessor> queueProcessorMap = new ConcurrentHashMap<>();
+    private final OneToOneConcurrentArrayQueue<Supplier<StaticEventProcessor>> toStartList = new OneToOneConcurrentArrayQueue<>(128);
+    private final List<EventQueueToEventProcessor> queueReadersToAdd = new ArrayList<>();
+    private final FluxtionServer fluxtionServer;
+    private final DeadWheelScheduler scheduler;
+    private final Service<com.fluxtion.server.service.scheduler.SchedulerService> schedulerService;
+
+    public ComposingEventProcessorAgent(String roleName,
+                                        com.fluxtion.server.subscription.EventFlowManager eventFlowManager,
+                                        FluxtionServer fluxtionServer,
+                                        DeadWheelScheduler scheduler,
+                                        ConcurrentHashMap<String, Service<?>> registeredServices) {
+        super(roleName, scheduler);
+        this.eventFlowManager = eventFlowManager;
+        this.fluxtionServer = fluxtionServer;
+        this.scheduler = scheduler;
+        this.registeredServices = registeredServices;
+        this.schedulerService = new Service<>(scheduler, SchedulerService.class);
+    }
+
+    public void addEventFeedConsumer(Supplier<StaticEventProcessor> initFunction) {
+        toStartList.add(initFunction);
+    }
+
+    @Override
+    public void onStart() {
+        log.info("onStart");
+        super.onStart();
+    }
+
+    @Override
+    public int doWork() throws Exception {
+        toStartList.drain(init -> {
+            StaticEventProcessor eventProcessor = init.get();
+            com.fluxtion.server.subscription.EventFlowManager.setCurrentProcessor(eventProcessor);
+            eventProcessor.registerService(schedulerService);
+            registeredServices.values().forEach(eventProcessor::registerService);
+            eventProcessor.addEventFeed(this);
+            if (eventProcessor instanceof Lifecycle) {
+                ((Lifecycle) eventProcessor).start();
+            }
+            EventFlowManager.removeCurrentProcessor();
+        });
+
+        if (!queueReadersToAdd.isEmpty()) {
+            if (tryAdd(queueReadersToAdd.get(0))) {
+                queueReadersToAdd.remove(0);
+            }
+        }
+
+        return super.doWork();
+    }
+
+    @Override
+    public void onClose() {
+        log.info("onClose");
+        super.onClose();
+    }
+
+    @Override
+    public void registerSubscriber(StaticEventProcessor subscriber) {
+        log.info("registerSubscriber:" + subscriber);
+    }
+
+    @Override
+    public void subscribe(StaticEventProcessor subscriber, com.fluxtion.server.subscription.EventSubscriptionKey<?> subscriptionKey) {
+        Objects.requireNonNull(subscriber, "subscriber is null");
+        Objects.requireNonNull(subscriptionKey, "subscriptionKey is null");
+        log.info("subscribe subscriptionKey:" + subscriptionKey + " subscriber:" + subscriber);
+
+        EventQueueToEventProcessor eventQueueToEventProcessor = queueProcessorMap.get(subscriptionKey);
+
+        if (eventQueueToEventProcessor == null) {
+            eventQueueToEventProcessor = eventFlowManager.getMappingAgent(subscriptionKey, this);
+            queueProcessorMap.put(subscriptionKey, eventQueueToEventProcessor);
+            queueReadersToAdd.add(eventQueueToEventProcessor);
+            log.info("added new subscribe subscriptionKey:" + subscriptionKey + " subscriber:" + subscriber);
+        }
+
+        eventQueueToEventProcessor.registerProcessor(subscriber);
+        eventFlowManager.subscribe(subscriptionKey);
+    }
+
+    @Override
+    public void unSubscribe(StaticEventProcessor subscriber, EventSubscriptionKey<?> subscriptionKey) {
+        if (queueProcessorMap.containsKey(subscriptionKey)) {
+            EventQueueToEventProcessor eventQueueToEventProcessor = queueProcessorMap.get(subscriptionKey);
+            if (eventQueueToEventProcessor.deregisterProcessor(subscriber) == 0) {
+                log.info("EventQueueToEventProcessor listener count = 0, removing subscription:" + subscriptionKey);
+                queueProcessorMap.remove(subscriptionKey);
+                eventFlowManager.unSubscribe(subscriptionKey);
+            }
+        }
+    }
+
+    @Override
+    public void removeAllSubscriptions(StaticEventProcessor subscriber) {
+
+    }
+}
