+++ src/main/java/com/fluxtion/server/subscription/EventSourceKey.java
@@ -1,0 +1,16 @@
+/*
+ * SPDX-FileCopyrightText: © 2024 Gregory Higgins <greg.higgins@v12technology.com>
+ * SPDX-License-Identifier: AGPL-3.0-only
+ *
+ */
+
+package com.fluxtion.server.subscription;
+
+import com.fluxtion.runtime.annotations.feature.Experimental;
+import lombok.Value;
+
+@Experimental
+@Value
+public class EventSourceKey<T> {
+    String sourceName;
+}
